Splitting out grpc table provider to its own file for code cleanliness
